+++ src/components/pie-player/pie-player.tsx
@@ -1,5 +1,7 @@
 import {SessionChangedEvent} from "@pie-framework/pie-player-events";
-import * as mr from "@pie-lib/math-rendering";
+import {
+  _dll_pie_lib__pie_toolbox_math_rendering_accessible
+} from "@pie-lib/pie-toolbox-math-rendering-module/module";
 import {
   Component,
   Element,
@@ -13,6 +15,7 @@ import {
 } from "@stencil/core";
 import {
   AdvancedItemConfig,
+  Env,
   ItemConfig,
   ItemSession,
   PieContent,
@@ -114,6 +117,11 @@ export class Player {
    */
   @Prop() addCorrectResponse: boolean = false;
 
+  /**
+   * In evaluate mode, add a bottom border to visually separate each item in the case of a multi-item
+   */
+  @Prop() showBottomBorder: boolean = false;
+
   @Watch("addCorrectResponse")
   watchAddCorrectResponse(newValue, oldValue) {
     if (newValue !== oldValue) {
@@ -130,7 +138,7 @@ export class Player {
    * Describes runtime environment for the player.
    *
    */
-  @Prop() env: Object = {mode: "gather", role: "student"};
+  @Prop() env: Env = {mode: "gather", role: "student"};
 
   /**
    * Indicates if player running in the context of a PIE hosting system.
@@ -383,8 +391,24 @@ export class Player {
 
   private renderMath() {
     setTimeout(() => {
-      mr.renderMath(this.el);
+      _dll_pie_lib__pie_toolbox_math_rendering_accessible.renderMath(this.el);
     }, 50);
+  }
+
+  private addBottomBorder(tags: string[]) {
+    if (!Array.isArray(tags)) {
+      return;
+    }
+
+    tags.forEach(tag => {
+      const elems = document.querySelectorAll(`${tag}`)
+
+      for (const elem of elems) {
+        if (elem && elem instanceof HTMLElement) {
+          elem.classList.add('evaluate-bottom-border');
+        }
+      }
+    });
   }
 
   async afterRender() {
@@ -392,6 +416,11 @@ export class Player {
       if (this.elementsLoaded) {
         this.updateModels();
         this.renderMath();
+
+        if (this.showBottomBorder && this.env.mode === 'evaluate') {
+          const pieTags = this.pieContentModel.elements && Object.keys(this.pieContentModel.elements)
+          this.addBottomBorder(pieTags);
+        }
       } else {
         const elements = Object.keys(this.pieContentModel.elements).map(el => ({
           name: el,